requirements.txt
tensorflow
numpy
pandas
pillow
scikit-learn
cleverhans

